requirements.txt
altgraph==0.17.3
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
comm==0.1.3
comtypes==1.1.14
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
idna==3.4
ipykernel==6.23.1
ipython==8.13.2
jaraco.context==4.3.0
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
matplotlib-inline==0.1.6
more-itertools==9.1.0
MouseInfo==0.1.3
mysql-connector-python==8.1.0
nest-asyncio==1.5.6
packaging==23.1
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
protobuf==4.21.12
psutil==5.9.5
pure-eval==0.2.2
PyAudio==0.2.13
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
Pygments==2.15.1
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
pyjokes==0.6.0
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pyqt5-tools==5.15.9.3.3
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
PyQt6-sip==13.5.2
pyqt6-tools==6.4.2.3.3
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
python-dotenv==1.0.0
pyttsx3==2.90
pytweening==1.0.7
pywin32==306
pywin32-ctypes==0.2.0
pyzmq==25.0.2
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
SpeechRecognition==3.10.0
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
urllib3==2.0.2
vboxapi==1.0
wcwidth==0.2.6
wikipedia==1.4.0
wolframalpha==5.0.0
xmltodict==0.13.0
